update util access and update hotslop for nav button
