+++ packages/Calendar/src/util.js
@@ -1,7 +1,5 @@
-import { getTheme } from '@sharingapples/theme';
-
-const theme = getTheme();
-const calendarTheme = theme.onCalendar || theme;
+import { StyleSheet } from 'react-native';
+import { primaryColor } from './theme';
 
 /* eslint-disable no-bitwise */
 
@@ -9,12 +7,24 @@ export const DAY_DIFF = 86400 * 1000;
 export const SEVEN_DAYS = 7;
 export const SINGLE_DAY = 1;
 
-const borderStyles = Array(16).fill(null).map((n, idx) => ({
-  borderTopColor: idx & 1 ? 'transparent' : calendarTheme.primary,
-  borderRightColor: idx & 2 ? 'transparent' : calendarTheme.primary,
-  borderBottomColor: idx & 4 ? 'transparent' : calendarTheme.primary,
-  borderLeftColor: idx & 8 ? 'transparent' : calendarTheme.primary,
-}));
+export const WEEK_DAYS = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
+export const ALL_MONTHS = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sept', 'Oct', 'Nov', 'Dec'];
+
+
+const borderStyles = Array(16).fill(null).map((n, idx) => StyleSheet.create({
+  border: {
+    alignItems: 'center',
+    paddingHorizontal: 3,
+    width: `${100 / 7}%`,
+    paddingVertical: 3,
+    borderWidth: StyleSheet.hairlineWidth,
+    margin: -StyleSheet.hairlineWidth,
+    borderTopColor: idx & 1 ? 'transparent' : primaryColor,
+    borderRightColor: idx & 2 ? 'transparent' : primaryColor,
+    borderBottomColor: idx & 4 ? 'transparent' : primaryColor,
+    borderLeftColor: idx & 8 ? 'transparent' : primaryColor,
+  },
+}).border);
 
 
 const cache = {};
@@ -38,6 +48,11 @@ export function generateYears(year) {
   return Array(16).fill(0).map((m, idx) => year + idx);
 }
 
+export function getMonthCount({ width }, date) {
+  const value = Array.isArray(date) ? date[0] : date;
+  return width > 560
+    ? [value, new Date(value.getFullYear(), value.getMonth() + 1)] : [value];
+}
 
 function getDateString(date, diff) {
   return new Date(date.getFullYear(), date.getMonth(), date.getDate() + diff).toDateString();
@@ -45,18 +60,18 @@ function getDateString(date, diff) {
 
 export function getDateBorderStyle(date, selectedDates) {
   if (!selectedDates) {
-    return undefined;
+    return borderStyles[15];
   }
 
   if (isDate(selectedDates)) {
-    return selectedDates.toDateString() === date.toDateString() ? borderStyles[0] : undefined;
+    return selectedDates.toDateString() === date.toDateString() ? borderStyles[0] : borderStyles[15];
   }
 
   const allDates = memoize(selectedDates);
 
   const currentDateString = date.toDateString();
   if (!allDates[currentDateString]) {
-    return undefined;
+    return borderStyles[15];
   }
 
   const prevWeek = getDateString(date, -SEVEN_DAYS);
